pyproject.toml
[project]
name = "ps3-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "torch>=2.7.1",
    "python-dotenv>=1.1.1",
    "pdf2image>=1.17.0",
    "langchain-community>=0.3.27",
    "google-genai>=1.31.0",
    "google-generativeai>=0.8.5",
]
